fix: build errors return type

+++ src/RawPOSIX/src/safeposix/dispatcher.rs
@@ -309,7 +309,7 @@ pub fn lind_syscall_api(
             // Turn off PROT_EXEC for non-code pages
             let prot = prot & !PROT_EXEC;
         
-            interface::mmap_handler(cageid, addr, len, prot, flags, fd, off)
+            interface::mmap_handler(cageid, addr, len, prot, flags, fd, off) as i32
         }
 
         PREAD_SYSCALL => {
